Updates to use N-PMMoV only for sewage data
This update trades off a bit of noise reduction (gained by averaging
N+S data) for a bit of potential future-proofing.  The success of
recent variants has been driven by spike mutations, and the "S" is
short for "Spike."  This update moves over to using only the N
dataset, in anticipation of possibly losing the ability to effectively
surveil for spike data.

It also fixes a bug when plotting the raw sewage datapoints: it was
previously plotting the dataset as points, instead of plotting the raw
data as points.